Switch to s16 free list arrays

+++ src/game_engine_stuff/sprite_stuff/sprite_allocator_class.hpp
@@ -33,9 +33,9 @@ protected:		// variables
 	sa_free_list_backend the_sa_free_list_backend;
 	
 public:		// functions
-	sprite_allocator( int* the_sa_free_list_backend_array, 
+	sprite_allocator( s16* the_sa_free_list_backend_array, 
 		sprite* s_the_array, u32 s_size );
-	sprite_allocator( int* the_sa_free_list_backend_array,
+	sprite_allocator( s16* the_sa_free_list_backend_array,
 		const array_helper<sprite>& s_allocatable_sprite_arr );
 	
 	////void* allocate_sprite() __attribute__((_iwram_code,hot));
